pyproject.toml
[project]
name = "satellite-image-analyzer"
version = "2.0.0"
description = "Система анализа спутниковых снимков (GUI + алгоритмы детекции/улучшения SAR)"
readme = "programm/README.md"
requires-python = ">=3.9"
authors = [{ name = "Diplom Project" }]

dependencies = [
  "PyQt5==5.15.10",
  "numpy>=1.26.0",
  "opencv-python>=4.9.0",
  "Pillow>=9.0.0",
  "requests>=2.25.0",
  "PyYAML>=6.0",
  "torch>=2.0.0",
  "torchvision>=0.15.0",
]

[project.optional-dependencies]
dev = ["pytest", "black", "flake8", "mypy"]
onnx = ["onnxruntime>=1.16.0"]

[project.urls]
homepage = "https://example.com"

[tool.pdm]
distribution = true

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["main*"]
exclude = ["programm*", "research_algorithm*"]


